+++ public/js/ajax/user/home.ajax.js
@@ -1,0 +1,2 @@
+// Create an instance of Notyf
+const notyf = new Notyf();