+++ resources/views/admin/home.blade.php
@@ -8,6 +8,6 @@
   @endslot
 
   @slot('customJS')
-
+  <script src="{{ asset('js/ajax/admin/home.ajax.js')}}" defer></script>
   @endslot
 </x-admin-layout>